+++ src/main/java/org/goafabric/calleeservice/Application.java
@@ -6,11 +6,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 
-
-/**
- * Created by amautsch on 26.06.2015.
- */
-
 @SpringBootApplication
 public class Application {
 
@@ -22,5 +17,5 @@ public class Application {
     public CommandLineRunner init(ApplicationContext context) {
         return args -> {if ((args.length > 0) && ("-check-integrity".equals(args[0]))) {SpringApplication.exit(context, () -> 0);}};
     }
-    
+
 }