+++ doc/release-notes.md
@@ -1,3 +1,6 @@
+# 3.2.10
+- spring security readded
+
 # 3.2.5
 - release plugin added
 - spring security removed